cleaned up example
